+++ src/main/java/com/mallangs/domain/community/entity/Like.java
@@ -1,6 +1,6 @@
 package com.mallangs.domain.community.entity;
 
-import com.mallangs.domain.member.Member;
+import com.mallangs.domain.member.entity.Member;
 import com.mallangs.global.common.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.*;
