+++ src/main/java/com/mallangs/domain/notification/entity/Notification.java
@@ -1,6 +1,5 @@
 package com.mallangs.domain.notification.entity;
 
-import com.mallangs.domain.member.Member;
 import com.mallangs.global.common.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.*;
